final main.go and import cleanup
